+++ messages/de/activities.php
@@ -1,5 +1,0 @@
-<?php
-return array (
-    'Bookmarked' => 'Lesezeichen',
-    'Whenever someone bookmarked something (e.g. a post or comment).' => 'Wann immer irgendjemand etwas mit einem Lesezeichen versetzt (z.B. Beitrag oder Kommentar).',
-);