+++ messages/nb-NO/forms.php
@@ -17,6 +17,16 @@
  * NOTE: this file must be saved in UTF-8 encoding.
  */
 return [
-    'Bookmarked' => '',
-    'Whenever someone bookmarked something (e.g. a post or comment).' => '',
+    '<strong>Bookmark</strong> module configuration' => '',
+    'Back to modules' => '',
+    'Bookmark view settings' => '',
+    'Here you can change the view settings for bookmarks.' => '',
+    'Icon color' => '',
+    'Select color' => '',
+    'Show Bookmark-Count in braces.' => '',
+    'Show an icon for each entry.' => '',
+    'Show global Bookmark-View in full width.' => '',
+    'Show pinned content first.' => '',
+    'Sort order' => '',
+    'Keep bookmarks private and do not send notifications.' => '',
 ];